Reorganize utils module into a directory
